Add files via upload
Foo Corporation needs a program to calculate how much to pay their hourly employees. The US Department of Labor 
requires that employees get paid time and a half for any hours over 40 that they work in a single week. For example, if an 
employee works 45 hours, they get 5 hours of overtime, at 1.5 times their base pay. The State of Massachusetts requires 
that hourly employees be paid at least $8.00 an hour. Foo Corp requires that an employee not work more than 60 hours in 
a week.